build: Move src/ directory to eos-updater/
This completes our move to per-program/library subdirectories, and fixes
the build from the previous commit.


https://phabricator.endlessm.com/T15816
