+++ eos-updater/apply.c
@@ -20,13 +20,11 @@
  *  - Vivek Dasmohapatra <vivek@etla.org>
  */
 
-#include "eos-updater-apply.h"
-#include "eos-updater-data.h"
-#include "eos-updater-object.h"
-
+#include <eos-updater/apply.h>
+#include <eos-updater/data.h>
+#include <eos-updater/object.h>
 #include <libeos-updater-util/types.h>
 #include <libeos-updater-util/util.h>
-
 #include <ostree.h>
 
 /* Closure containing the data for the apply worker thread. The
