+++ eos-updater/com.endlessm.Updater.conf
