+++ eos-updater/com.endlessm.Updater.xml
