+++ eos-updater/live-boot.h
@@ -1,6 +1,6 @@
 /* -*- mode: C; c-file-style: "gnu"; indent-tabs-mode: nil; -*-
  *
- * Copyright © 2013 Collabora Ltd.
+ * Copyright © 2016 Endless Mobile
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -17,23 +17,20 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  *
  * Authors:
- *  - Vivek Dasmohapatra <vivek@etla.org>
+ *  - Will Thompson <wjt@endlessm.com>
  */
 
 #pragma once
 
-#include "eos-updater-generated.h"
-
-#include <gio/gio.h>
+#include <eos-updater/dbus.h>
+#include <libeos-updater-util/util.h>
+#include <ostree.h>
 
 G_BEGIN_DECLS
 
-gboolean handle_poll (EosUpdater            *updater,
-                      GDBusMethodInvocation *call,
-                      gpointer               user_data);
-gboolean handle_poll_volume (EosUpdater            *updater,
-                             GDBusMethodInvocation *call,
-                             const gchar           *path,
-                             gpointer               user_data);
+gboolean is_installed_system (GError **error);
+gboolean handle_on_live_boot (EosUpdater            *updater,
+                              GDBusMethodInvocation *call,
+                              gpointer               user_data);
 
 G_END_DECLS
