+++ eos-updater/poll-common.h
@@ -22,14 +22,11 @@
 
 #pragma once
 
-#include "eos-updater-data.h"
-
-#include <libeos-updater-util/refcounted.h>
-
-#include <ostree.h>
-
+#include <eos-updater/data.h>
 #include <glib.h>
 #include <gio/gio.h>
+#include <libeos-updater-util/refcounted.h>
+#include <ostree.h>
 
 G_BEGIN_DECLS
